refactor: rename ml services
