+++ videoprocessor/services/__init__.py
@@ -1,7 +1,7 @@
 from .video_file_manager import VideoFileManager
 from .base_ml_service import BaseMLService
-from .mock_ml_predict import MockMLService
-from .send_video_to_ml import APIMLService
+from .mock_ml_service import MockMLService
+from .api_ml_service import APIMLService
 from .ml_service import get_ml_service, ml_service
 from .video_upload_service import VideoUploadService
 from .video_processing_service import VideoProcessingService
